requirements.txt
python-telegram-bot==21.6
google-generativeai==0.7.0
pdfplumber==0.11.4
python-docx==1.1.2
flask==3.0.3
twilio==9.2.3
requests==2.32.3
python-dotenv==1.0.1
